requirements.txt
fastapi==0.128.7
uvicorn==0.40.0
motor==3.7.1
python-dotenv==1.0.1
openpyxl==3.1.2
